Remove exported ports.
The "ports" fields in docker-compose.yml define which ports are published
to the outside world. For testing purposes, it's not necessary to expose
any of these ports. It's particularly nice to avoid publishing ports,
because Docker's publishing mechanism actually bypasses host-based
firewalls like ufw, so people may be surprised to find that running
Boulder in Docker for testing purposes exposes these ports.