NonceService: Take deletion out of critical path
Once an instance of NonceService is full (which it is in the steady
state), every nonce redemption takes a lock then scans the map of used
nonces looking for the smallest value. That value is removed, and the
low water mark of acceptable nonce counter values is increased.

That lock has become a bottleneck, since every POST request has to wait
on the same lock, and we do nontrivial work (3ms) while holding the
lock.

In this change, set a target size for the map of used nonces instead of
a max size. This allows us to hold the lock for a much smaller amount of
work, just adding the nonce counter value to the used map. Then, in the
background, a cull routine works to shrink the map back down to the
target size.

During heavy traffic, when the map grows above the target, the cull
routine will go to work. This means that incoming requests will be
competing for the lock with `cull`. Fortunately, cull will unlock the
mutex every 3ms, allowing the waiting requests to acquire it. Go does
not guarantee mutex fairness, but it should be reasonable based on

https://go-review.googlesource.com/c/go/+/34310
https://github.com/golang/go/issues/13086

In the worst case, cull will hog the mutex until it succeeds in bringing
the map size back to its target. In this scenario, incoming requests are
blocked on acquiring the lock, so the map isn't growing, and cull will
eventually achieve its goal. However, there's an O(n^2) problem here,
since the bigger the map gets, the longer the linear scan takes.