Revocation: only require 1 valid authz for acct revocation.
Proving valid authorization for one domain name present in a certificate
is now considered sufficient authorization to revoke the entire
certificate when features.OneAuthzRevocation is enabled. When it is not
enabled valid authorizations are required for *all* names present in
a certificate in order to be authorized to revoke the certificate.